Fin de WorldService.getMap()
