+++ src/maps/TerrainVariables.ts
@@ -2,4 +2,6 @@ const WATER_HEIGHT = 0.4
 const MOUNTAIN_HEIGHT = 0.8
 const TILE_SIZE = 16
 
-export {WATER_HEIGHT, MOUNTAIN_HEIGHT, TILE_SIZE}
+const IS_HEIGHT_WALKABLE = (height: number) => height > WATER_HEIGHT && height < MOUNTAIN_HEIGHT
+
+export { WATER_HEIGHT, MOUNTAIN_HEIGHT, TILE_SIZE, IS_HEIGHT_WALKABLE }
